requirements.txt
flask==3.0.3
python-binance==1.0.19
websocket-client==1.8.0
pandas==2.2.2
